First approach to password encryption
