+++ package.json
@@ -14,6 +14,7 @@
     "url": "git+https://github.com/vctr90/expenses-tracker-api.git"
   },
   "dependencies": {
+    "bcrypt": "^3.0.4",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.9",
     "dotenv": "^6.2.0",
